Remove legacy utility methods
