+++ tests/test_trainer.py
@@ -9,7 +9,6 @@ import example_project_2 as ex2
 
 from aind_behavior_curriculum import (
     GRADUATED,
-    INIT_STAGE,
     Stage,
     Trainer,
     TrainerState,
@@ -92,12 +91,14 @@ class TrainerTests(unittest.TestCase):
         # Create single-stage curriculum
         taskA = ex.TaskA(task_parameters=ex.TaskAParameters())
         stageA = Stage(name="StageA", task=taskA)
-        stageA.add_policy_transition(INIT_STAGE, ex.stageA_policyB, ex.t1_10)
-        stageA.add_policy_transition(INIT_STAGE, ex.stageA_policyA, ex.t1_5)
+        stageA.add_policy_transition(
+            ex.INIT_STAGE, ex.stageA_policyB, ex.t1_10
+        )
+        stageA.add_policy_transition(ex.INIT_STAGE, ex.stageA_policyA, ex.t1_5)
         stageA.add_policy_transition(
             ex.stageA_policyA, ex.stageA_policyB, ex.t1_10
         )
-        stageA.set_start_policies(INIT_STAGE)
+        stageA.set_start_policies(ex.INIT_STAGE)
 
         stageAA = stageA.model_copy(deep=True)
         stageAA.set_task_parameters(
@@ -114,9 +115,9 @@ class TrainerTests(unittest.TestCase):
 
         # Associate mice with curriculum
         tr = ex.ExampleTrainer()
-        tr.register_subject(0, curr, stageA, INIT_STAGE)
-        tr.register_subject(1, curr, stageA, INIT_STAGE)
-        tr.register_subject(2, curr, stageA, INIT_STAGE)
+        tr.register_subject(0, curr, stageA, ex.INIT_STAGE)
+        tr.register_subject(1, curr, stageA, ex.INIT_STAGE)
+        tr.register_subject(2, curr, stageA, ex.INIT_STAGE)
 
         # Move mice through curriculum by updating metrics
         ex.MICE_METRICS[0] = ex.ExampleMetrics(theta_1=0)
@@ -138,10 +139,10 @@ class TrainerTests(unittest.TestCase):
         # Validate mouse histories
         M0 = [
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
                 stage=stageAA, active_policies=[ex.stageA_policyA]
@@ -153,7 +154,7 @@ class TrainerTests(unittest.TestCase):
 
         M1 = [
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
                 stage=stageAA, active_policies=[ex.stageA_policyA]
@@ -168,7 +169,7 @@ class TrainerTests(unittest.TestCase):
 
         M2 = [
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
                 stage=stageAAA, active_policies=[ex.stageA_policyB]
@@ -196,8 +197,8 @@ class TrainerTests(unittest.TestCase):
         # StageA and StageB picked out from observation
         stageA = curr.see_stages()[0]
         stageB = curr.see_stages()[2]
-        stageA.set_start_policies([INIT_STAGE])
-        stageB.set_start_policies([INIT_STAGE])
+        stageA.set_start_policies([ex.INIT_STAGE])
+        stageB.set_start_policies([ex.INIT_STAGE])
 
         stageAA = stageA.model_copy(deep=True)
         stageAA.set_task_parameters(
@@ -248,13 +249,13 @@ class TrainerTests(unittest.TestCase):
         # Validate mouse histories
         M0 = [
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
                 stage=stageAA, active_policies=[ex.stageA_policyA]
             ),
             trainer.create_trainer_state(
-                stage=stageB, active_policies=[INIT_STAGE]
+                stage=stageB, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
                 stage=stageBB, active_policies=[ex.stageB_policyA]
@@ -264,13 +265,13 @@ class TrainerTests(unittest.TestCase):
 
         M1 = [
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
                 stage=stageAAA, active_policies=[ex.stageA_policyB]
             ),
             trainer.create_trainer_state(
-                stage=stageB, active_policies=[INIT_STAGE]
+                stage=stageB, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
                 stage=stageBBB, active_policies=[ex.stageB_policyB]
@@ -280,7 +281,7 @@ class TrainerTests(unittest.TestCase):
 
         M2 = [
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
                 stage=stageAAA, active_policies=[ex.stageA_policyB]
@@ -325,7 +326,7 @@ class TrainerTests(unittest.TestCase):
         )
 
         tr = ex.ExampleTrainer()
-        tr.register_subject(0, curr, stageA, INIT_STAGE)
+        tr.register_subject(0, curr, stageA, ex.INIT_STAGE)
 
         # Override API
         tr.override_subject_status(
@@ -341,7 +342,7 @@ class TrainerTests(unittest.TestCase):
         # Validate mouse history
         M0 = [
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
                 stage=stageBB, active_policies=[ex.stageB_policyA]
@@ -365,9 +366,9 @@ class TrainerTests(unittest.TestCase):
         taskA = ex.TaskA(task_parameters=ex.TaskAParameters())
         taskB = ex.TaskB(task_parameters=ex.TaskBParameters())
         stageA = Stage(name="StageA", task=taskA)
-        stageA.set_start_policies(INIT_STAGE, append_non_existing=True)
+        stageA.set_start_policies(ex.INIT_STAGE, append_non_existing=True)
         stageB = Stage(name="StageB", task=taskB)
-        stageB.set_start_policies(INIT_STAGE, append_non_existing=True)
+        stageB.set_start_policies(ex.INIT_STAGE, append_non_existing=True)
 
         curr = ex.MyCurriculum(name="My Curriculum")
         curr.add_stage(stageA)
@@ -375,8 +376,8 @@ class TrainerTests(unittest.TestCase):
 
         # Associate mice with curriculum
         tr = ex.ExampleTrainer()
-        tr.register_subject(0, curr, stageA, INIT_STAGE)
-        tr.register_subject(1, curr, stageB, INIT_STAGE)
+        tr.register_subject(0, curr, stageA, ex.INIT_STAGE)
+        tr.register_subject(1, curr, stageB, ex.INIT_STAGE)
 
         # Regular evaluation logs same entry
         ex.MICE_METRICS[0] = ex.ExampleMetrics(theta_1=10, theta_2=10)
@@ -401,31 +402,31 @@ class TrainerTests(unittest.TestCase):
         # Validate mouse history
         M0 = [
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
-                stage=stageB, active_policies=[INIT_STAGE]
+                stage=stageB, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
         ]
 
         M1 = [
             trainer.create_trainer_state(
-                stage=stageB, active_policies=[INIT_STAGE]
+                stage=stageB, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
-                stage=stageB, active_policies=[INIT_STAGE]
+                stage=stageB, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
-                stage=stageB, active_policies=[INIT_STAGE]
+                stage=stageB, active_policies=[ex.INIT_STAGE]
             ),
         ]
 
@@ -443,10 +444,10 @@ class TrainerTests(unittest.TestCase):
         taskA = ex.TaskA(task_parameters=ex.TaskAParameters())
         stageA = Stage(name="StageA", task=taskA)
 
-        stageA.add_policy(INIT_STAGE)
+        stageA.add_policy(ex.INIT_STAGE)
         stageA.add_policy(ex.stageA_policyA)
         stageA.add_policy(ex.stageA_policyB)
-        stageA.set_start_policies(INIT_STAGE)
+        stageA.set_start_policies(ex.INIT_STAGE)
 
         stageAA = stageA.model_copy(deep=True)
         stageAA.set_task_parameters(
@@ -462,8 +463,8 @@ class TrainerTests(unittest.TestCase):
 
         # Associate mice with curriculum
         tr = ex.ExampleTrainer()
-        tr.register_subject(0, curr, stageA, INIT_STAGE)
-        tr.register_subject(1, curr, stageA, INIT_STAGE)
+        tr.register_subject(0, curr, stageA, ex.INIT_STAGE)
+        tr.register_subject(1, curr, stageA, ex.INIT_STAGE)
 
         # Regular evaluation logs same entry
         ex.MICE_METRICS[0] = ex.ExampleMetrics(theta_1=10, theta_2=10)
@@ -488,10 +489,10 @@ class TrainerTests(unittest.TestCase):
         # Validate mouse history
         M0 = [
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
                 stage=stageA, active_policies=[ex.stageA_policyA]
@@ -503,10 +504,10 @@ class TrainerTests(unittest.TestCase):
 
         M1 = [
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
-                stage=stageA, active_policies=[INIT_STAGE]
+                stage=stageA, active_policies=[ex.INIT_STAGE]
             ),
             trainer.create_trainer_state(
                 stage=stageA, active_policies=[ex.stageA_policyB]
@@ -781,7 +782,7 @@ class TrainerTests(unittest.TestCase):
         )
 
         tr = ex.ExampleTrainer()
-        tr.register_subject(0, curr, stageA, INIT_STAGE)
+        tr.register_subject(0, curr, stageA, ex.INIT_STAGE)
 
         ex.MICE_METRICS[0] = ex.ExampleMetrics(theta_1=8)
         tr.evaluate_subjects()
@@ -796,7 +797,7 @@ class TrainerTests(unittest.TestCase):
             trainer.create_trainer_state(
                 stage=stageA,
                 is_on_curriculum=True,
-                active_policies=[INIT_STAGE],
+                active_policies=[ex.INIT_STAGE],
             ),
             trainer.create_trainer_state(
                 stage=stageAA,
@@ -810,7 +811,7 @@ class TrainerTests(unittest.TestCase):
                 stage=None, is_on_curriculum=False, active_policies=None
             ),
             trainer.create_trainer_state(
-                stage=stageB, active_policies=[INIT_STAGE]
+                stage=stageB, active_policies=[ex.INIT_STAGE]
             ),
         ]
         self.assertEqual(tr.subject_history[0], M0)
@@ -834,7 +835,7 @@ class TrainerStateTests(unittest.TestCase):
     def setUp(self):
         self.curr = ex.construct_curriculum()
         stageA = self.curr.see_stages()[0]
-        active_policies = [INIT_STAGE]
+        active_policies = [ex.INIT_STAGE]
 
         self.state_from_trainer_state = TrainerState(
             curriculum=self.curr,
